- changed file structure; - set count of threads in decompose_v2_par
